[linstor-migrator] add stage3
